backend: rework `queue::push` after v2

+++ backend/windmill-worker/src/worker.rs
@@ -146,7 +146,7 @@ use crate::mssql_executor::do_mssql;
 use crate::bigquery_executor::do_bigquery;
 
 #[cfg(feature = "benchmark")]
-use windmill_common::bench::{benchmark_init, BenchmarkInfo, BenchmarkIter};
+use crate::bench::{benchmark_init, BenchmarkInfo, BenchmarkIter};
 
 use windmill_common::add_time;
 
@@ -1050,9 +1050,9 @@ pub async fn run_worker(
     let is_dedicated_worker: bool = WORKER_CONFIG.read().await.dedicated_worker.is_some();
 
     #[cfg(feature = "benchmark")]
-    let benchmark_jobs: i32 = std::env::var("BENCHMARK_JOBS")
+    let benchmark_jobs: usize = std::env::var("BENCHMARK_JOBS")
         .unwrap_or("5000".to_string())
-        .parse::<i32>()
+        .parse::<usize>()
         .unwrap();
 
     #[cfg(feature = "benchmark")]
